rename cloud backup to keyless backup for consistency in naming
