+++ src/keylessBackup/SetUpKeylessBackup.tsx
@@ -11,8 +11,8 @@ import Card from 'src/components/Card'
 import Colors from 'src/styles/colors'
 import variables from 'src/styles/variables'
 import colors from 'src/styles/colors'
-import EnvelopeIcon from 'src/cloudBackup/EnvelopeIcon'
-import SmartphoneIcon from 'src/cloudBackup/SmartphoneIcon'
+import EnvelopeIcon from 'src/keylessBackup/EnvelopeIcon'
+import SmartphoneIcon from 'src/keylessBackup/SmartphoneIcon'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { KeylessBackupEvents } from 'src/analytics/Events'
