+++ src/navigator/Navigator.tsx
@@ -122,7 +122,7 @@ import VerificationStartScreen from 'src/verify/VerificationStartScreen'
 import WalletConnectSessionsScreen from 'src/walletConnect/screens/Sessions'
 import WalletConnectRequest from 'src/walletConnect/screens/WalletConnectRequest'
 import WebViewScreen from 'src/webview/WebViewScreen'
-import SetUpKeylessBackup from 'src/cloudBackup/SetUpKeylessBackup'
+import SetUpKeylessBackup from 'src/keylessBackup/SetUpKeylessBackup'
 
 const TAG = 'Navigator'
